Replace existing top-level data with escalated context data
We need a way to escalate a "service" from the context, but the context
is always set. The existing way of escalating preserves existing values
and does not replace anything. Instead, we should back up the original
data in the context, prefixed by an underscore, then replace the
top-level data.